+++ src/main/java/com/l02gr01/escape/model/history/event/Loss.java
@@ -3,7 +3,7 @@ package com.l02gr01.escape.model.history.event;
 public class Loss extends Event{
 
     public int keysgrabbed;
-    public Loss(String name, int time, int finallevel, int keysgrabbed) {
+    public Loss(String name, long time, int finallevel, int keysgrabbed) {
         super(name, time, finallevel);
         this.keysgrabbed = keysgrabbed;
     }
